move general to datasetaggregator
